Using SMC

SMC Main Function

You can use SMC to estimate any Bayesian model. This requires (1) parameters and their associated prior distributions (2) data (3) a likliehood function. These three ingredients are the only inputs into the smc driver.

smc(loglikelihood::Function, parameters::ParameterVector, data::Matrix) function.

Let's look at an example. First, we'll make a model. We need the ModelConstructors package to do that.

using ModelConstructors
ols_model = GenericModel()

Model 
no. parameters:             0
description:
 A generic model: You can put parameters, states, shocks, observables, pseudo-observables, spec, subpsec, settings inside of me

Next, assign our intercept and coefficient parameters to the model.

reg <= parameter(:α1, 0., (-1e5, 1e5), (-1e5, 1e5), Untransformed(), Normal(0, 10), fixed = false)
reg <= parameter(:β1, 0., (-1e5, 1e5), (-1e5, 1e5), Untransformed(), Normal(0, 10), fixed = false)

The first argument is the name of the parameter, the second is its initial value, the third and fourther are bounds on the parameter (they should always be the same). If the sampler draws a value outside of these bounds, it'll try again. The fifth argument is whether you would like to perform a transformation on the parameter when using it (ignore this for now–it's for more advanced users). The sixth argument is the prior, $N(0, 10)$, and last argument says the parameters aren't fixed parameters (i.e. we're trying to estimate them!)

We'll make some artifical data to test our model

β = 1.
α = 1.
y = β*X + α

100-element Array{Float64,1}:
 1.7982768624815304
 1.8706763462995302
 1.6611300317024342
 1.279692124647822 
 1.1163322701458438
 1.9220390217503973
 1.2144784620817473
 1.3498425765900104
 1.8941287983018167
 1.097760213045384 
 ⋮                 
 1.0266587997750434
 1.447363286258172 
 1.2938889744335087
 1.7418306223962174
 1.5805577525308265
 1.295438513730291 
 1.055235600862271 
 1.469619166037717 
 1.7100962913432323

We need a log-likelihood function! Note that it's important you define a log-likelihood function rather than a likelihood function.

function likelihood_fnct(p, d)
    α = p[1]
    β = p[2]
    Σ = 1
    det_Σ = det(Σ)
    inv_Σ = inv(Σ)
    term1 = -N / 2 * log(2 * π) - 1 /2 * log(det_Σ)
    logprob = 0.
    errors = d[:, 1] .- α .- β .* d[:, 2]
    for t in 1:size(d,1)
        logprob += term1 - 1/2 * dot(errors, inv_Σ * errors)
    end
    return logprob
end

likelihood_fnct (generic function with 1 method)

And that's it! Now let's run SMC.

smc(likelihood_fnct, reg.parameters, data, n_parts = 10, use_fixed_schedule = false, tempering_target = 0.97)

SMC.smc — Method.

function smc(loglikelihood::Function, parameters::ParameterVector{U}, data::Matrix{S};
             kwargs...) where {S<:AbstractFloat, U<:Number}

Arguments:

loglikelihood::Function: Log-likelihood function of model being estimated. Takes parameters and data as arguments.

parameters::ParameterVector{U}: Model parameter vector, which stores parameter values, prior dists, and bounds.

data: A matrix or dataframe containing the time series of the observables used in the calculation of the posterior/loglikelihood

old_data: A matrix containing the time series of observables of previous data (with data being the new data) for the purposes of a time tempered estimation (that is, using the posterior draws from a previous estimation as the initial set of draws for an estimation with new data)

Keyword Arguments:

verbose::Symbol: Desired frequency of function progress messages printed to standard out.

- `:none`: No status updates will be reported.
- `:low`: Status updates for SMC initialization and recursion will be included.
- `:high`: Status updates for every iteration of SMC is output, which includes
the mean and standard deviation of each parameter draw after each iteration,
as well as calculated acceptance rate, ESS, and number of times resampled.

parallel::Bool: Flag for running algorithm in parallel.

n_parts::Int: Number of particles.

n_blocks::Int: Number of parameter blocks in mutation step.

n_mh_steps::Int: Number of Metropolis Hastings steps to attempt during the mutation step.

λ::S: The 'bending coefficient' λ in Φ(n) = (n/N(Φ))^λ

n_Φ::Int: Number of stages in the tempering schedule.

resampling_method::Symbol: Which resampling method to use.

:systematic: Will use sytematic resampling.

:multinomial: Will use multinomial resampling.

:polyalgo: Samples using a polyalgorithm.

threshold_ratio::S: Threshold s.t. particles will be resampled when the population drops below threshold * N.

c::S: Scaling factor for covariance of the particles. Controls size of steps in mutation step.

α::S: The mixture proportion for the mutation step's proposal distribution.

target::S: The initial target acceptance rate for new particles during mutation.

use_chand_recursion::Bool: Flag for using Chandrasekhar Recursions in Kalman filter.

use_fixed_schedule::Bool: Flag for whether or not to use a fixed tempering (ϕ) schedule.

tempering_target::S: Coefficient of the sample size metric to be targeted when solving for an endogenous ϕ.

old_data::Matrix{S}: data from vintage of last SMC estimation.

old_cloud::Cloud: associated cloud borne of old data in previous SMC estimation.

old_vintage::String: String for vintage date of old data

smc_iteration::Int: The iteration index for the number of times SMC has been run on the same data vintage. Primarily for numerical accuracy/testing purposes.

run_test::Bool: Flag for when testing accuracy of program

filestring_addl::Vector{String}: Additional file string extension for loading old cloud.

continue_intermediate::Bool: Flag to indicate whether one is continuing SMC from an intermediate stage.

intermediate_stage_start::Int: Intermediate stage at which one wishes to begin the estimation.

save_intermediate::Bool: Flag for whether one wants to save intermediate Cloud objects

intermediate_stage_increment::Int: Save Clouds at every increment (1 = each stage, 10 = every 10th stage, etc.)

Outputs

cloud: The Cloud object containing all of the information about the parameter values from the sample, their respective log-likelihoods, the ESS schedule, tempering schedule etc., which is saved in the saveroot.

Overview

Sequential Monte Carlo can be used in lieu of Random Walk Metropolis Hastings to generate parameter samples from high-dimensional parameter spaces using sequentially constructed proposal densities to be used in iterative importance sampling.

This implementation is based on Edward Herbst and Frank Schorfheide's 2014 paper 'Sequential Monte Carlo Sampling for DSGE Models' and the code accompanying their book 'Bayesian Estimation of DSGE Models'.

SMC is broken up into three main steps:

Correction: Reweight the particles from stage n-1 by defining incremental weights, which gradually "temper in" the loglikelihood function p(Y|θ)^(ϕn - ϕn-1) into the normalized particle weights.

Selection: Resample the particles if the distribution of particles begins to degenerate, according to a tolerance level for the ESS.

Mutation: Propagate particles {θ(i), W(n)} via N(MH) steps of a Metropolis Hastings algorithm.

source