Helper Functions

Helper Functions

SMC.Cloud — Method.

function Cloud(n_params::Int, n_parts::Int)

Easier constructor for Cloud, which initializes the weights to be equal, and everything else in the particle object etc. to be empty.

source

Missing docstring.

Missing docstring for SMC.get_weights(c::Cloud). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SMC.get_vals(c::Cloud; transpose::Bool = true). Check Documenter's build log for details.

Missing docstring.

Missing docstring for prior(parameters::ParameterVector{T}) where {T<:Number}. Check Documenter's build log for details.

SMC.solve_adaptive_ϕ — Method.

`function solve_adaptive_ϕ(cloud::Cloud, proposed_fixed_schedule::Vector{Float64},
                           i::Int64, j::Int64, ϕ_prop::Float64, ϕ_n1::Float64,
                           tempering_target::Float64, resampled_last_period::Bool)`

Solves for next Φ. Returns ϕn, resampledlastperiod, j, ϕprop.

source

SMC.mvnormal_mixture_draw — Method.

`mvnormal_mixture_draw(θ_old::Vector{T}, d_prop::Distribution;
                       c::T = 1.0, α::T = 1.0) where T<:AbstractFloat`

Create a DegenerateMvNormal distribution object, d, from a parameter vector, p, and a standard deviation matrix (obtained from SVD), σ.

Generate a draw from the mixture distribution of:

A DegenerateMvNormal centered at θ_old with the standard deviation matrix σ, scaled by cc^2 and with mixture proportion α.

A DegenerateMvNormal centered at the same mean, but with a standard deviation matrix of the diagonal entries of σ scaled by cc^2 with mixture proportion (1 - α)/2.

A DegenerateMvNormal with the same standard deviation matrix σ but centered at the new proposed mean, θ_prop, scaled by cc^2, and with mixture proportion (1 - α)/2.

If no θ_prop is given, but an α is specified, then the mixture will consist of α of the standard distribution and (1 - α) of the diagonalized distribution.

Arguments

θ_old::Vector{T}: The mean of the desired distribution

σ::Matrix{T}: The standard deviation matrix of the desired distribution

Keyword Arguments

cc::T: The standard deviation matrix scaling factor

α::T: The mixing proportion

θ_prop::Vector{T}: The proposed parameter vector to be used as part of the mixture distribution, set by default to be the weighted mean of the particles, prior to mutation.

Outputs

θ_new::Vector{T}: The draw from the mixture distribution to be used as the MH proposed step

source

SMC.compute_ESS — Method.

function `compute_ESS(loglh::Vector{T}, current_weights::Vector{T}, ϕ_n::T, ϕ_n1::T;
                     old_loglh::Vector{T} = zeros(length(loglh))) where {T<:AbstractFloat}`

Compute ESS given log likelihood, current weights, ϕn, ϕ{n-1}, and old log likelihood.

source

SMC.generate_free_blocks — Method.

`generate_free_blocks(n_free_para::Int64, n_blocks::Int64)`

Return a Vector{Vector{Int64}} where each internal Vector{Int64} contains a subset of the range 1:nfreepara of randomly permuted indices. This is used to index out random blocks of free parameters from the covariance matrix for the mutation step.

source

SMC.generate_all_blocks — Method.

`generate_all_blocks(blocks_free::Vector{Vector{Int64}}, free_para_inds::Vector{Int64})`

Return a Vector{Vector{Int64}} where each internal Vector{Int64} contains indices corresponding to those in blocks_free but mapping to 1:n_para (as opposed to 1:n_free_para). These blocks are used to reconstruct the particle vector by inserting the mutated free parameters into the size n_para, particle vector, which also contains fixed parameters.

source

SMC.mutation — Method.

mutation(loglikelihood::Function, parameters::ParameterVector{U},
         data::Matrix{S}, p::Vector{S}, d_μ::Vector{S}, d_Σ::Matrix{S},
         blocks_free::Vector{Vector{Int}}, blocks_all::Vector{Vector{Int}},
         ϕ_n::S, ϕ_n1::S; c::S = 1., α::S = 1., n_mh_steps::Int = 1,
         old_data::T = T(undef, size(data, 1), 0)) where {S<:AbstractFloat,
                                                          T<:AbstractMatrix, U<:Number}

Execute one proposed move of the Metropolis-Hastings algorithm for a given parameter

Arguments:

loglikelihood::Function: Likelihood function of model being estimated.

parameters::ParameterVector{U}: Model parameter vector, which stores parameter values, prior dists, and bounds

data::Matrix{Float64}: Matrix of data

p::Vector{Float64}: Initial particle value

d::Distribution: A distribution with μ = the weighted mean, and Σ = the weighted variance/covariance matrix

blocks_free::Vector{Vector{Int64}}: A vector of index blocks, where the indices in each block corresponds to the ordering of free parameters only (e.g. all the indices will be ∈ 1:nfreeparameters)

blocks_all::Vector{Vector{Int64}}: A vector of index blocks, where the indices in each block corresponds to the ordering of all parameters (e.g. all the indices will be in ∈ 1:n_para, free and fixed)

ϕ_n::Float64: The current tempering factor

ϕ_n1::Float64: The previous tempering factor

Keyword Arguments:

c::Float64: The scaling parameter for the proposed covariance matrix

α::Float64: The mixing proportion

n_mh_steps::Int: Number of Metropolis Hastings steps to attempt

old_data::Matrix{Float64}: The matrix of old data to be used in calculating the oldloglh, oldlogpost in time tempering

Outputs:

p::Vector{Float64}: An updated particle containing updated parameter values, log-likelihood, prior, and acceptance indicator.

source

Missing docstring.

Missing docstring for one_draw(loglikelihood::Function, parameters::ParameterVector{U}, data::Matrix{Float64}) where {U<:Number}. Check Documenter's build log for details.

SMC.initial_draw! — Method.

function initial_draw!(loglikelihood::Function, parameters::ParameterVector{U},
                       data::Matrix{Float64}, c::Cloud; parallel::Bool = false) where {U<:Number}

Draw from a general starting distribution (set by default to be from the prior) to initialize the SMC algorithm. Returns a tuple (logpost, loglh) and modifies the particle objects in the particle cloud in place.

source

Missing docstring.

Missing docstring for draw_likelihood(loglikelihood::Function, parameters::ParameterVector{U}, data::Matrix{Float64}, draw::Vector{Float64}) where {U<:Number}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for initialize_likelihoods!(loglikelihood::Function, parameters::ParameterVector{U}, data::Matrix{Float64}, c::Cloud; parallel::Bool = false) where {U<:Number}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for initialize_cloud_settings!(cloud::Cloud; tempered_update::Bool = false, n_parts::Int = 5_000, n_Φ::Int = 300, c::S = 0.5, accept::S = 0.25) where {S<:AbstractFloat}. Check Documenter's build log for details.